backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
alembic==1.13.1
python-dotenv==1.0.0
tiktoken>=0.5.0
